requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.5.4
tensorflow==1.14.0
termcolor==1.1.0
wrapt==1.12.1
